[test] Split cell/row editing tests
